+++ CHANGELOG.md
@@ -9,6 +9,16 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v5.6x.0
+
+### Fixed
+
+- Fix handle model not found thrown inside a custom mutation.
+
+### Added
+
+- Add a new error handler `ModelNotFoundErrorHandler` to handle models not found correctly, before this had the category internal server error.
+
 ## v5.63.0
 
 ### Added